backend/requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
DateTime==5.2
Deprecated==1.2.14
firebase-admin==6.2.0
Flask==3.0.0
Flask-CORS==4.0.0
frozenlist==1.4.0
gcloud==0.18.3
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.12.0
google-cloud-storage==2.11.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jwcrypto==1.5.0
MarkupSafe==2.1.3
msgpack==1.0.7
multidict==6.0.4
oauth2client==4.1.3
proto-plus==1.22.3
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3.post1
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
twilio==8.9.1
uritemplate==4.1.1
urllib3==1.26.17
uuid==1.30
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
zope.interface==6.1
Flask-RESTful==0.3.10
app.config==1.0.1
boto==2.49.0
pyotp==2.9.0
pytest==7.4.2
Pillow==9.3.0
pytest==7.4.2
Pillow==9.3.0